requirements.txt
docopt>=0.6.2
gitdb>=0.6.4
GitPython>=2.0.7
Jinja2>=2.8
MarkupSafe>=0.23
smmap>=0.9.0
